requirements.txt
aiogram
sqlalchemy
redis
python-dotenv
rich
GitPython
aiosqlite